Make infinite feed list unique for each new feed batch
This is caused because of the changes I made in the FeedList component
where I have a local state for given feed list, so that whenever the category
gets changed, the changes should be reflected on the list without having to
refetch the list again.

This calls for a global state, where I should do this and not scatter this
around couple of layers deep and few components.

That one was tough to find :/

I need e2e tests, pronto.